+++ src/main/java/lotto/LottoApplication.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.controller.LottoController;
 import lotto.domain.LottoStore;
 import lotto.service.WinningLotto;
